Slim the diff to lite
