+++ test/manual.js
@@ -1,7 +1,7 @@
 // bun --inspect-brk test/manual.js      --no-foo bar --foo2 bar2
 import argMate, { argInfo } from '../src/argMateLite.js';
 
-import precompileConfig from '../src/configPrep.js';
+import precompileConfig from '../src/compileConfig.js';
 
 debugger;
